Improve POR for exceptions.
Keep track of thread masking states and only consider a 'throwTo' as
dependent if the action being thrown at can be interrupted. This halves
the execution time of the testsuite on my linode (~190s to ~90s).

Requires dpor-0.2, as thread IDs must be known when stepping the state.

Closes #32, although I feel like there might be yet more gains possible.

-- unconfirmed hunch follows --

I feel like there might be a bug lurking here, but I haven't been able
to exhibit one.

My intuition is that because throwing an exception which can't interrupt
WILL block and so enable other threads, this might have issues in
combination with preemption bounding, although the testsuite still
passes.