Implement thread priorities.
In scheduling, rather than choose from the entire 'canTry' list, only
the highest-priority threads are chosen from. There is no upper bound on
priorities. 0 is the default priority of every thread.

When a thread is created, a random priority-assignment function is
called to assign it an initial priority. When a random
priority-reassignment predicate returns true, every thread is assigned a
new priority.