Expose the discard logic through Test.DejaFu
New functions: dejafuDiscard and dejafuDiscardIO

No dejafusDiscard and dejafusDiscardIO, because I think the
combination of discarding + sharing traces would be very confusing.
Also, the dejafus* functions are for when you want to trade space for
time, whereas the *Discard* functions are for the opposite.  So it
doesn't make much sense.