Set up doctests
Only adds one doctest so far, but gets the machinery working.  To
solve the problem of dejafu's fancy output, this checks a
"DEJAFU_DOCTEST" env var: when set, the output is simplified.

Doctests run in a separate Travis job (rather than the existing "test"
job) because the version of doctest in LTS-6 doesn't like CPP.  I feel
that as long as they're working for the latest LTS, all is fine.