Add dependency function property tests
The idea is that, if the dependency function is correct, then any
permutation of independent actions should give the same result.

In practice it's a little more complex, as the dependency function has
some special cases which are only valid because of how it is used.  So
an "independence" function which we can use to shuffle any trace
after-the-fact has to work around those cases.