Prune commits before barriers if there are no other buffered writes
It's only sound to remove a commit if there are no other buffered
writes for that `IORef`.  Say we have a `IORef` with two buffered
writes, followed by a memory barrier, we get (amongst others) these
two trace fragments:

    C1  ->  barrier        = C2 wins
    C2  ->  barrier        = C1 wins

If we collapse these two to just `barrier`, then we will only get one
of the results.