Replace dontCheck/subconcurrency with new Program class
The 'Program' is the new formulation of dejafu unit tests.  A
'Program' is one of three types:

- A 'ConcT', which is as before.

- A 'WithSetup', which corresponds to 'dontCheck'.

- A 'WithSetupAndTeardown', which corresponds to 'subconcurrency'.

This more new formulation makes it impossible to nest 'withSetup' (the
replacement for 'dontCheck') or 'withSetupAndTeardown' (the
half-replacement for 'subconcurrency'), by as these functions take a
'ConcT' as their argument and produce a 'Program WithSetup' or
'Program WithSetupAndTeardown'.

The testing functions have all been generalised to work with this
'Program' type.