Implement dejafu in terms of a "MonadDejaFu" class
dejafu needs a few things which can't be implemented purely:

- mutable references
- bound threads

Mutable references used to come from the MonadRef class, but with the
introduction of bound threads in 1.0.0.0, it seemed easiest to use a
MonadConc as the underlying monad.  This also allowed simplifying the
API a lot, and was generally a good move.

But it lost purity.  By *requiring* bound threads be supported by the
base monad, ST was out.

Making bound threads optional (by moving rtsSupportsBoundThreads into
MonadConc and introducing an additional layer of abstraction in
dejafu) makes it possible to recover purity.  Pure implementations
just won't support bound threads.