Add Concourse pipeline
Putting the tasks in separate files would be nice, but that doesn't
play so nicely with the path-restricted git resources.  Such a
resource pulls the git version *which changed the path*, not the
latest, so I'd need to pull in a second, unrestricted, input to get
the task files from.

That makes the pipeline look kind of awkward in the UI, so I decided
to just inline the definitions here.